Refactor: Split cJSON_Number into cJSON_Double
We barely use the double code (only in reading in failover log and
when we parse the configuration file). Let's handle double's in a
special form there to avoid information loss for integers.

Reviewed-on: http://review.couchbase.org/90055
Tested-by: Build Bot <build@couchbase.com>

+++ daemon/connection.cc
@@ -220,7 +220,7 @@ unique_cJSON_ptr Connection::toJSON() const {
     if (socketDescriptor == INVALID_SOCKET) {
         cJSON_AddStringToObject(obj, "socket", "disconnected");
     } else {
-        cJSON_AddNumberToObject(obj, "socket", (double)socketDescriptor);
+        cJSON_AddNumberToObject(obj, "socket", socketDescriptor);
         cJSON_AddStringToObject(obj, "protocol", "memcached");
         cJSON_AddStringToObject(obj, "peername", getPeername().c_str());
         cJSON_AddStringToObject(obj, "sockname", getSockname().c_str());
